+++ src/shared/enums/db/mongodb-errors.enum.ts
@@ -1,0 +1,7 @@
+export enum MongooseErrorCodes {
+  UniqueConstraintFail = 11000
+}
+
+export enum MongooseErrors {
+  MongoServerError = 'MongoServerError'
+}