+++ src/shared/enums/messages/error-messages.enum.ts
@@ -4,6 +4,8 @@ export enum ErrorMessages {
   GetFail = 'Unable to retrieve data from DB!',
   UpdateFail = 'Unable to update data in DB!',
   DeleteFail = 'Unable to delete entry from DB!',
+  DuplicateEntryFail = 'User already exists!',
+  PasswordMismatchFail = 'Passwords must match!',
   Generic = 'Something went wrong!',
 	NotFound = 'Unable to find the requested resource!',
 }